Update contao44 to 4.4.4.
Here is release announce:

Contao version 4.4.4 is available. The bugfix release fixes several issues
including a problem with the back end referrer management. In addition, the
back end theme has been adjusted so the submit buttons are shown at the end of
the form again instead of at the end of the page.