postgresql-postgis2: Work around an upstream configure bug
Upstream has very complicated ifdefs that try to define UNIX
everywhere but windows, except that it fails to include NetBSD.  Then,
it uses UNIX to avoid using the non-standard stricmp and instead use
the POSIX-conforming strcasecmp.

For pkgsrc, rip out an entire screenful of complexity and just define
UNIX if not defined.  Resolves build on NetBSD 7, and probably other places.