Initial import of R-curl v3.0 from CRAN.
The curl() and curl_download() functions provide highly configurable
drop-in replacements for base url() and download.file() with better
performance, support for encryption (https, ftps), gzip compression,
authentication, and other 'libcurl' goodies. The core of the package
implements a framework for performing fully customized requests where
data can be processed either in memory, on disk, or streaming via the
callback or connection interfaces.